% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_boundary.R
\name{catalog_boundary}
\alias{catalog_boundary}
\title{Computes the polygons that encloses the points}
\usage{
catalog_boundary(ctg, concavity = 5, lengthThreshold = 5, simplify = 5)
}
\arguments{
\item{ctg}{A LAScatalog}

\item{concavity}{numeric a relative measure of concavity. 1 results in a relatively detailed shape,
Infinity results in a convex hull. You can use values lower than 1, but they can produce pretty crazy
shapes.}

\item{lengthThreshold}{numeric. when a segment length is under this threshold, it stops being
considered for further detalization. Higher values result in simpler shapes.}
}
\value{
A LAScatalog with true boundaries
}
\description{
Computes the polygons that encloses the points. It reads all the file one by one and computes a
concave hull using the concaveman. When all the hulls are computed it updates the LAScatalog to
set the true polygons instead of the bounding boxes.
}
\section{Supported processing options}{

Supported processing options for more details see the
\link[lidR:LAScatalog-class]{LAScatalog engine documentation}:
\itemize{
\item chunk size: Not suported, it processes by file.
\item chunk buffer: Not suported, it processes by file with no buffer.
\item chunk alignment: Not suported, it processes by file.
\item \strong{progress}: Displays a progress estimate.
\item output files: Not supported, it returns an R object
\item select: Not suported, it loads XYZ only.
\item \strong{filter}: Read only the points of interest.
}
}

\examples{
LASfile <- system.file("extdata", "Megaplot.laz", package="lidR")
ctg <- readLAScatalog(LASfile, filter = "-drop_z_below 2")
ctg2 <- catalog_boundary(ctg, 2, 2)
plot(ctg)
plot(ctg2, add = TRUE)
}
