% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_flightlines_z_alignment.R
\name{check_flightlines_z_alignment}
\alias{check_flightlines_z_alignment}
\title{Check if flightlines are aligned}
\usage{
check_flightlines_z_alignment(las, res = 3)
}
\arguments{
\item{las}{a LAS point cloud}
}
\description{
Check if flightlines are well aligned by comparing the average difference between the lowest points
of overlapping parts of the flightlines
}
