% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerlines-classify_transmissiontowers.R
\name{classify_transmissiontowers}
\alias{classify_transmissiontowers}
\title{Classify the transmission towers}
\usage{
classify_transmissiontowers(las, towers, dtm, threshold = 2)
}
\arguments{
\item{las}{An object of class LAS}

\item{towers}{SpatialPointsDataFrame returned by \link{find_transmissiontowers}.}

\item{dtm}{A RasterLayer. The digital terrain model is useful to find the bottom of the towers}

\item{threshold}{numeric. Height above ground. Points below this elevation are not classified
as transmission towers.}
}
\value{
A LAS object with an updated classification.
}
\description{
Attribute the class 15 to points that correspond to transmission towers using the positionning
of the towers given by \link{find_transmissiontowers}. Using the transmission towers types and their
orientation it computes the spatial box that emcompasses the towers and classify points within this
rectangle as 'transmission tower'.
}
\examples{
\donttest{
# A simple file with wires already clipped from 4 files + shapefile
# of the network
LASfile <- system.file("extdata", "wires.laz", package="lidRplugins")
wireshp <- system.file("extdata", "wires.shp", package="lidRplugins")
las <- readLAS(LASfile, select = "xyzc")
network <- raster::shapefile(wireshp)

# Ugly dtm because data is an Y
dtm <- grid_terrain(las, 2, tin())

towers <- find_transmissiontowers(las, network, dtm, "waist-type")
las <- classify_transmissiontowers(las, towers, dtm)

plot(las, color = "Classification")
}
}
\references{
Roussel J, Achim A, Auty D. 2021. Classification of high-voltage power line structures in low density
ALS data acquired over broad non-urban areas. PeerJ Computer Science 7:e672 https://doi.org/10.7717/peerj-cs.672
}
\seealso{
Other electrical network: 
\code{\link{classify_wires}()},
\code{\link{get_tower_spec}()},
\code{\link{track_wires}()}
}
\concept{electrical network}
