% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerlines-classify_wires.R
\name{classify_wires}
\alias{classify_wires}
\title{Classify wire}
\usage{
classify_wires(las, wires, dtm)
}
\arguments{
\item{las}{An object of class LAS.}

\item{wires}{A \code{SpatialPointsDataFrame} returned by \link{track_wires}}

\item{dtm}{A RasterLayer. The digital terrain model is need to get the relative elevations}
}
\description{
Classify wires using the output of \link{track_wires}. Using wire profiles and some knowledge
on the geometry of the towers it creates a cloth below the wires and classifies everything that
is above this cloth will be classifies a wire exept the transmission towers. This the transmission
tower classification must be performed first.
}
\examples{
\donttest{
# A simple file with wires already clipped from 4 files + shapefile
# of the network
LASfile <- system.file("extdata", "wires.laz", package="lidRplugins")
wireshp <- system.file("extdata", "wires.shp", package="lidRplugins")
dtmtif  <- system.file("extdata", "wire-dtm.tif", package="lidRplugins")
las <- readLAS(LASfile, select = "xyzc")
network <- sf::st_read(wireshp)
dtm <- raster::raster(dtmtif)

towers <- find_transmissiontowers(las, network, dtm, "waist-type")
las <- classify_transmissiontowers(las, towers, dtm)
wires <- track_wires(towers, network, dtm, "waist-type")
las <- classify_wires(las, wires, dtm)

plot(las, color = "Classification")
}
}
\references{
Roussel J, Achim A, Auty D. 2021. Classification of high-voltage power line structures in low density
ALS data acquired over broad non-urban areas. PeerJ Computer Science 7:e672 https://doi.org/10.7717/peerj-cs.672
}
\seealso{
Other electrical network: 
\code{\link{classify_transmissiontowers}()},
\code{\link{get_tower_spec}()},
\code{\link{track_wires}()}
}
\concept{electrical network}
