% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_transmissiontowers.R
\name{find_transmissiontowers}
\alias{find_transmissiontowers}
\title{Individual transmission tower detection}
\usage{
find_transmissiontowers(
  las,
  powerline,
  dtm,
  type = c("waist-type", "double-circuit"),
  buffer = 125,
  debug = FALSE
)
}
\arguments{
\item{las}{An object of class LAS with absolute elevations or a LAScatalog.}

\item{powerline}{A \code{SpatialLines*} that map the electrical network accurately}

\item{dtm}{\code{RasterLayer}. Because the algorithm relies on absolute elevation a DTM is
requirered to compute the relative elevations.}

\item{type}{character. One of "waist-type", "waist-type-small", "double-circuit" according to
\href{http://www.hydroquebec.com/learning/transport/types-pylones.html}{Hydro-Quebec}. Can also
be a list with custom specifications. See \link{get_tower_spec}.}

\item{buffer}{numeric. The \code{SpatialLines*} will be buffered internally to catch the powerlines
and the transmission towers. The buffer must ensure to catch all the powerlines.}

\item{debug}{logical. Plot the different steps of the algorithm so one can try to figure out what
is going wrong.}
}
\value{
A \code{SpatialPointDataFrame} with several attributes. \code{Z} the elevation of the tower,
\code{dtm} the elevation of the bottom of the tower aligned with the top, \code{theta} the angle
of the tower with the x axis in radian, \code{ux, uy} the directional vectors, \code{deflection}
tells if a given tower is on a deflection (deflection towers are found twice by design) and
\code{type} the type name.
}
\description{
Individual transmission tower detection function that find the positions of the transmission
towers. The method is supervised by a map of the electric network and the tower types.
}
\examples{
\donttest{
# A simple file with wires already clipped from 4 files + shapefile
# of the network
LASfile <- system.file("extdata", "wires.laz", package="lidRplugins")
wireshp <- system.file("extdata", "wires.shp", package="lidRplugins")
las <- readLAS(LASfile, select = "xyzc")
network <- shapefile(wireshp)

dtm <- grid_terrain(las, 2, tin())

towers <- find_transmissiontowers(las, network, dtm, "waist-type")

plot(las@header)
plot(towers, add = TRUE, col = towers$deflection + 1)
arrows(
   towers@coords[,1],
   towers@coords[,2],
   towers@coords[,1] + 100 * towers$ux,
   towers@coords[,2] + 100 * towers$uy,
   length = 0.05,
   col = towers$deflection + 1)

x = plot(las)
add_treetops3d(x, towers, radius = 5)
}
}
\concept{electric network}
