% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_flightlines_z_alignment.R
\name{flightlines_z_misalignment_matrix}
\alias{flightlines_z_misalignment_matrix}
\title{Check if flightlines are aligned}
\usage{
flightlines_z_misalignment_matrix(las, res = 5)
}
\arguments{
\item{las}{a LAS point cloud}
}
\value{
An adjacency matrix with the offsets between overlapping flightlines
}
\description{
Check if flightlines are well aligned by comparing the average difference between the lowest points
of overlapping parts of the flightlines
}
