% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerlines-find_transmissiontowers.R
\name{get_tower_spec}
\alias{get_tower_spec}
\title{Get specification of a tower}
\usage{
get_tower_spec(type)
}
\arguments{
\item{type}{one of \code{"waist-type"} \code{"double-circuit"} or \code{"waist-type-small"} or a
\code{list}}
}
\value{
A list
}
\description{
Return the specifications of a given tower type i.e. size range, width range, tension, number of
wire. If \code{type} is contains the specifications of a non supported type it checks the validity
of the specification
}
\examples{
specs = get_tower_spec("waist-type")

# Create new specs
new_specs = specs
new_specs$width = c(20, 25)

# Validate this specs
 get_tower_spec(new_specs)
}
\seealso{
Other electrical network: 
\code{\link{classify_transmissiontowers}()},
\code{\link{classify_wires}()},
\code{\link{track_wires}()}
}
\concept{electrical network}
