% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lascoplanar.r
\name{lascoplanar}
\alias{lascoplanar}
\title{Test if points are part of a neighborhood that is approximately coplanar}
\usage{
lascoplanar(las, k = 8, th1 = 25, th2 = 6)
}
\arguments{
\item{las}{an object of class LAS}

\item{k}{interger. The number of k-nearest neighbors.}

\item{th1}{numeric. The threshold to be applied to the smallest eigenvalue.}

\item{th2}{numeric. The threshold to be applied to the second smallest eigenvalue.}
}
\value{
A LAS object with a new column names \code{Coplanar} (or \code{Colinear} if \code{th2 = 0})
that indicates those points that are part of a neighborhood that is approximately coplanar/colinear
 (TRUE) or not (FALSE).
}
\description{
An approximate coplanarity test. For each point it looks for the k-nearest neighbors. It computes the
eigenvalues of the covariance matrix. The points that meet the following criteria are labeled as
approximately coplanar:\cr
\deqn{λ_2 > (th_1*λ_1) && (th_2*λ_2) > λ_3}
\eqn{λ_1, λ_2, λ_3} are the eigenvalues of a neighborhood of points (defined by k-nearest neighbors)
in ascending order.\cr\cr
If \code{th2 = 0} the function turns to be a colinearity test. The points that meet the following criteria are labeled as
approximately colinear:\cr
\deqn{th_1*λ_2 < λ_3 && (th_1*λ_1) < λ_3}
}
