% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LayerStacking.r
\name{LayerStacking}
\alias{LayerStacking}
\title{Individual Tree Detection Algorithm}
\usage{
LayerStacking(
  start = 0.5,
  res = 1,
  ws1 = 3,
  ws2 = 1.5,
  buf_size = 0.5,
  hardwood = FALSE,
  hmin = 2
)
}
\arguments{
\item{start}{scalar The point cloud is horizontally layered at 1-m intervals starting at 'start'
meters above the ground. Default is 0.5 (page 18)}

\item{res}{scalar Resolution of the CHM computed with a point-to-raster approach. Default is 1 (page 19).}

\item{ws1}{scalar Windows radius of the first local maxima use to detected tree tops on the CHM.
Default is 3 (page 19)}

\item{ws2}{scalar Windows radius of the second local maxima used to detected tree tops on the overlap
map. Default is 1.5 (page 20)}

\item{buf_size}{scalar Buffer size placed around each point to build a polygonal buffer around each
cluster (figure 1c page 20). Default is 0.5 (page 18)}

\item{hardwood}{logical. In dense conifer stands with little penetration to the center of the tree,
additional weight on the overlap map is given to clusters (page 20). Default is FALSE}

\item{hmin}{scalar. Point below this threshold cannot initiate a new tree.}
}
\description{
This function is made to be used in \link[lidR:tree_detection]{tree_detection}. It implements the
LayerStacking algorithm for tree detection based on Ayrey et al (2017) (see references).
This function implements only the fisrt part of the method i.e. the detection of the trees.
}
\references{
Ayrey, E., Fraver, S., Kershaw, J. A., Kenefic, L. S., Hayes, D., Weiskittel, A. R., & Roth, B. E.
(2017). Layer Stacking: A Novel Algorithm for Individual Forest Tree Segmentation from LiDAR Point
Clouds. Canadian Journal of Remote Sensing, 43(1), 16–27. https://doi.org/10.1080/07038992.2017.1252907
}
\seealso{
Other individual tree detection algorithms: 
\code{\link{lmfauto}()},
\code{\link{multichm}()},
\code{\link{ptrees}()}
}
\concept{individual tree detection algorithms}
