% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algo-lmfx.R
\name{lmfx}
\alias{lmfx}
\title{Individual Tree Detection Algorithm}
\usage{
lmfx(ws, hmin = 2, dist_2d = 3)
}
\arguments{
\item{ws}{numeric or function. Length or diameter of the moving window used to the detect the local
maxima in the unit of the input data (usually meters). If it is numeric a fixed windows size is used.
If it is a function, the function determines the size of the window at any given location on the canopy.
The function should take the height of a given pixel or points as its only argument and return the
desired size of the search window when centered on that pixel/point.}

\item{hmin}{numeric. Minimum height of a tree. Threshold below which a pixel or a point
cannot be a local maxima. Default 2.}

\item{dist_2d}{numeric. 2D distance threshold. A local maximum is considered a detected tree
if there is no detected tree within this 2D distance.}
}
\description{
This function is made to be used in \link[lidR:find_trees]{find_trees}. It implements an
experimental algorithms for tree detection based on a several ideas from the litterature. First it
select the highest points in each cell of a 1 m grid to reduce the amount of data and considerably
improve speed, then it performs a local maximum filter to find tree tops. To finish it applies the
filtering rule from \link{multichm} (step (b))
}
\examples{
LASfile <- system.file("extdata", "MixedConifer.laz", package="lidR")
las = readLAS(LASfile)

ttops = tree_detection(las, lmfx(ws = 3))

x = plot(las)
add_treetops3d(x, ttops)
}
